+++ bundles/org.openhab.binding.vallox/src/main/java/org/openhab/binding/vallox/internal/se/mapper/FanChannel.java
@@ -30,7 +30,7 @@ public class FanChannel extends ValloxChannel {
      *
      * @param variable channel as byte
      */
-    public FanChannel(byte variable) {
+    public FanChannel(int variable) {
         super(variable);
     }
