+++ bundles/org.openhab.binding.vallox/src/main/java/org/openhab/binding/vallox/internal/se/mapper/StringChannel.java
@@ -29,7 +29,7 @@ public class StringChannel extends ValloxChannel {
      *
      * @param variable channel as byte
      */
-    public StringChannel(byte variable) {
+    public StringChannel(int variable) {
         super(variable);
     }
